requirements.txt
catboost==1.1.1
pandas==1.3.3
scikit-learn==0.24.2
numpy==1.21.2
google-cloud-bigquery==2.28.1
google-cloud-storage==1.42.3
KFP